Revert "Refactor nav menu into common module"
